8150: Re-enable USB Host by default.
On the previous USB driver the USB Host mode was disabled which causes unpowered OTG to not work at all, but offers better better compatibility with chargers.

It was changed to Enabled by default since is more versatile with OTG working at the cost of only PD chargers to charge the device.